+++ test/ext/lenient_map_test.exs
@@ -1,0 +1,5 @@
+defmodule Ext.LenientMapTest do
+  use ExUnit.Case, async: true
+
+  doctest Ext.LenientMap
+end